+++ voltixApp/VoltixApp/Views/Keygen/PeerDiscoveryView.swift
@@ -5,7 +5,6 @@
 
 import OSLog
 import SwiftUI
-import UniformTypeIdentifiers
 
 struct PeerDiscoveryView: View {
     let tssType: TssType
@@ -14,7 +13,6 @@ struct PeerDiscoveryView: View {
     @StateObject var viewModel = KeygenPeerDiscoveryViewModel()
     @StateObject var participantDiscovery = ParticipantDiscovery(isKeygen: true)
     @State private var orientation = UIDevice.current.orientation
-    @State var showFileImporter = false
     
     let columns = [
         GridItem(.adaptive(minimum: 160)),
@@ -38,37 +36,6 @@ struct PeerDiscoveryView: View {
             }
             ToolbarItem(placement: .topBarTrailing) {
                 NavigationHelpButton()
-            }
-        }
-        .fileImporter(
-            isPresented: $showFileImporter,
-            allowedContentTypes: [UTType.data],
-            allowsMultipleSelection: false
-        ) { result in
-            
-            switch result {
-            case .success(let urls):
-                guard let url = urls.first else { return }
-                print(url)
-                let filename = url.lastPathComponent
-                
-                
-                let success = url.startAccessingSecurityScopedResource()
-                defer { url.stopAccessingSecurityScopedResource() }
-                
-                guard success else {
-                    return
-                }
-                
-                do {
-                    let fileContent = try String(contentsOf: url, encoding: .utf8)
-                } catch {
-                    print("Failed to read file: \(error.localizedDescription)")
-                }
-                
-                
-            case .failure(let error):
-                print("Error selecting file: \(error.localizedDescription)")
             }
         }
         .task {
@@ -257,20 +224,6 @@ struct PeerDiscoveryView: View {
                 .font(.body15MenloBold)
                 .multilineTextAlignment(.center)
                 .foregroundColor(.red)
-            
-            filePicker
-        }
-    }
-    
-    var filePicker: some View {
-        Button {
-            print("Toggling file importer")
-            showFileImporter.toggle()
-        } label: {
-            Image(systemName: "photo.on.rectangle.angled")
-                .font(.body16Menlo)
-                .foregroundColor(.neutral0)
-                .frame(width: 40, height: 40)
         }
     }
